tests and validation

+++ controller/network_api.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/coreos/etcd/clientv3/concurrency"
@@ -39,6 +40,9 @@ func (s *NetworkApi) UpdateNetwork(in *edgeproto.Network, cb edgeproto.NetworkAp
 	cur := edgeproto.Network{}
 	changed := 0
 
+	if clusterInstApi.UsesNetwork(&in.Key) {
+		return errors.New("Network cannot be modified while associated with a Cluster Instance")
+	}
 	err := s.sync.ApplySTMWait(ctx, func(stm concurrency.STM) error {
 		if !s.store.STMGet(stm, &in.Key, &cur) {
 			return in.Key.NotFoundError()
@@ -63,6 +67,9 @@ func (s *NetworkApi) DeleteNetwork(in *edgeproto.Network, cb edgeproto.NetworkAp
 	ctx := cb.Context()
 	if !s.cache.HasKey(&in.Key) {
 		return in.Key.NotFoundError()
+	}
+	if clusterInstApi.UsesNetwork(&in.Key) {
+		return errors.New("Network in use by Cluster Instance")
 	}
 	_, err := s.store.Delete(ctx, in, s.sync.syncWait)
 	return err
